+++ my-project/App.js
@@ -1,71 +1,123 @@
-import { NavigationContainer } from "@react-navigation/native";
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import React from "react";
+import { StatusBar } from 'expo-status-bar';
+import React, { useEffect ,useRef} from "react";
 import {
   StyleSheet, Button, View, SafeAreaView, Text, Alert, TextInput,
 } from 'react-native';
+import { updateMyCurrentLocation, getUserData } from './src/location';
 
-const Stack = createNativeStackNavigator();
+import MapView, {
+  PROVIDER_GOOGLE,
+  MAP_TYPES,
+  PROVIDER_DEFAULT,
+  UrlTile,
+  Marker,
+} from "react-native-maps";
+
+const INITIAL_REGION = {
+  latitude: 59.44089,
+  longitude: 24.72749,
+  latitudeDelta: 2.036923536294034,
+  longitudeDelta: 2.48705391138792,
+}
+
 
 function Separator() {
   return <View style={styles.separator} />;
 }
 
-const HomePage = ({ navigation }) => {
+function App() {
   const [currentUserName, onChangeCurrentUserName] = React.useState('Meng');
   const [trackedUserName, onChangeTrackedUserName] = React.useState('Daniel');
+  const mapRef = useRef(null)
+
+
+  // const [region, onChangeRegion] = React.useState({
+  //   region: {
+  //     latitude: 45.4310163,
+  //     longitude: 11.8796273,
+  //     latitudeDelta: 0.0922,
+  //     longitudeDelta: 0.0421,
+  //   },
+  // });
 
   return (
     <SafeAreaView style={styles.container}>
       <View>
         <Text style={styles.title}>
-          What's your name / ID?
+          The title and onPress handler are required. It is recommended to set accessibilityLabel to help make your app usable by everyone.
         </Text>
         <TextInput
           style={styles.input}
           onChangeText={onChangeCurrentUserName}
           value={currentUserName}
         />
+        <Button
+          title="Update My Current Location"
+          color="#fc4903"
+          onPress={() => updateMyCurrentLocation(currentUserName)}
+        />
       </View>
+      <Separator />
       <View>
         <Text style={styles.title}>
-          Who's name / ID are you tracking?
+          Adjust the color in a way that looks standard on each platform. On  iOS, the color prop controls the color of the text. On Android, the color adjusts the background color of the button.
+        </Text>
+        <Button
+          title="Write User Data"
+          disabled
+          color="#fc8003"
+          onPress={() => Alert.alert('User data not written')}
+        />
+      </View>
+      <Separator />
+      <View>
+        <Text style={styles.title}>
+          Read user data by userId
         </Text>
         <TextInput
           style={styles.input}
           onChangeText={onChangeTrackedUserName}
           value={trackedUserName}
         />
+
+        <Button
+          title="Track Your Love"
+          color="#fcc203"
+          onPress={() => getUserData(trackedUserName)}
+        />
       </View>
       <Separator />
-      <Button
-        title="Start Tracking!"
-        color="#fcc203"
-        onPress = {() => navigation.navigate('TrackerPage', { name: 'Jane' })}
-        //onPress={() => openTrackerView()}
-      />
-    </SafeAreaView>
-  )
-}
-
-const TrackerPage = ({ navigation, route }) => {
-  return <Text>This is {route.params.name}'s profile</Text>;
-};
-
-function App() {
-
-  return (
-    <NavigationContainer>
-      <Stack.Navigator>
-        <Stack.Screen
-          name = "HomePage"
-          component = {HomePage}
-          options = {{ title: "Welcome!" }}
+      {/* <View>
+        <Text style={styles.title}>
+          This layout strategy lets the title define the width of the button.
+        </Text>
+        <View style={styles.fixToText}>
+          <Button
+            title="Left button"
+            color="#d1ceb0"
+            onPress={() => Alert.alert('Left button')}
           />
-        <Stack.Screen name="TrackerPage" component={TrackerPage} />
-      </Stack.Navigator>
-      
-    </NavigationContainer>
+          <Button
+            title="Right button"
+            color="#d1ceb0"
+            onPress={() => Alert.alert('Right button pressed')}
+          />
+        </View>
+      </View> */}
+      <View style={styles.container}>
+        <MapView style={styles.map}       
+          provider={PROVIDER_GOOGLE}
+          ref={mapRef}
+          initialRegion={INITIAL_REGION}
+          showsMyLocationButton={true}
+          showsCompass={true}
+        >   
+        <Marker coordinate={{ latitude: 59.44089, longitude: 24.72749 }} title="My Location" />     
+        </MapView>
+      </View>
+
+      <StatusBar style="auto" />
+    </SafeAreaView>
   );
 }
 
@@ -85,10 +137,24 @@ const styles = StyleSheet.create({
     textAlign: 'center',
     marginVertical: 8,
   },
+  fixToText: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+  },
   separator: {
     marginVertical: 8,
     borderBottomColor: '#737373',
     borderBottomWidth: StyleSheet.hairlineWidth,
+  },
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  map: {
+    width: 350,
+    height: 300,
   },
 });
