+++ package.json
@@ -5,8 +5,10 @@
   "main": "index.js",
   "scripts": {
     "initial-setup": "npm i && cd my-project && npm i",
-    "start": "npm run run-web",
-    "run-web": "cd my-project && npm run web",
+    "start": "npm run web",
+    "web": "cd my-project && npm run start",
+    "android": "cd my-project && npm run android",
+    "ios": "cd my-project && npm run ios",
     "Daniel": "echo \"Daniel is the best man\"",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
